requirements.txt
Pillow>=10.0.0
pynput>=1.7.6
paho-mqtt>=2.0.0
openai>=1.0.0
PyYAML>=6.0
python-dotenv>=1.0.1
pyperclip>=1.8.2
requests>=2.28.0
Flask>=3.0.0
